refactor: reorganize build scripts and move adapter logic to lib directory
