+++ packages/cli/src/build.ts
